requirements.txt
langchain
langchain-community
openai
streamlit
sqlite3
python-dotenv